requirements.txt
streamlit
opencv-python-headless
streamlink
google-cloud-vision
google-cloud-pubsub
google-cloud-aiplatform
python-dotenv
